fix(subgoal): cast bf16 pipeline inputs to fp32 before encode
The streaming pipeline yields bfloat16 tensors but the frozen WM weights are
fp32, so jepa.encode(x) raised 'Input type BFloat16 and bias type float should
be the same'. Base/aux training avoided this via AMP autocast; the subgoal loop
has no autocast, so cast x and loc to float32 explicitly.

+++ examples/ac_video_jepa/maze/main_subgoal.py
@@ -55,8 +55,8 @@ def main():
     for epoch in range(epochs):
         t0 = time.time(); tot = 0.0; nb = 0
         for x, a, loc, _, _ in loader:
-            x = x.to(device, non_blocking=True)
-            loc = loc.to(device, non_blocking=True)             # [B,2,T] normalized positions
+            x = x.to(device, non_blocking=True).float()         # pipeline yields bf16; encoder weights are fp32
+            loc = loc.to(device, non_blocking=True).float()     # [B,2,T] normalized positions
             B, _, T = loc.shape
             with torch.no_grad():
                 z = jepa.encode(x)                              # [B,f,T,1,1]
